// lib/twitter/api.ts
export interface ContextAnnotation {
  /**
   *
   * @type {ContextAnnotationDomainFields}
   * @memberof ContextAnnotation
   */
  domain: ContextAnnotationDomainFields;
  /**
   *
   * @type {ContextAnnotationEntityFields}
   * @memberof ContextAnnotation
   */
  entity: ContextAnnotationEntityFields;
}
// ...
export interface Expansions {
  /**
   *
   * @type {Array<Media>}
   * @memberof Expansions
   */
  media?: Array<Media>;
  /**
   *
   * @type {Array<Place>}
   * @memberof Expansions
   */
  places?: Array<Place>;
  /**
   *
   * @type {Array<Poll>}
   * @memberof Expansions
   */
  polls?: Array<Poll>;
  /**
   *
   * @type {Array<Topic>}
   * @memberof Expansions
   */
  topics?: Array<Topic>;
  /**
   *
   * @type {Array<Tweet>}
   * @memberof Expansions
   */
  tweets?: Array<Tweet>;
  /**
   *
   * @type {Array<User>}
   * @memberof Expansions
   */
  users?: Array<User>;
}
// ...
export interface FullTextEntities {
  /**
   *
   * @type {Array<FullTextEntitiesAnnotationsInner>}
   * @memberof FullTextEntities
   */
  annotations?: Array<FullTextEntitiesAnnotationsInner>;
  /**
   *
   * @type {Array<CashtagEntity>}
   * @memberof FullTextEntities
   */
  cashtags?: Array<CashtagEntity>;
  /**
   *
   * @type {Array<HashtagEntity>}
   * @memberof FullTextEntities
   */
  hashtags?: Array<HashtagEntity>;
  /**
   *
   * @type {Array<MentionEntity>}
   * @memberof FullTextEntities
   */
  mentions?: Array<MentionEntity>;
  /**
   *
   * @type {Array<UrlEntity>}
   * @memberof FullTextEntities
   */
  urls?: Array<UrlEntity>;
}
// ...
export interface Media {
  /**
   * The height of the media in pixels.
   * @type {number}
   * @memberof Media
   */
  height?: number;
  /**
   * The Media Key identifier for this attachment.
   * @type {string}
   * @memberof Media
   */
  media_key?: string;
  /**
   *
   * @type {string}
   * @memberof Media
   */
  type: string;
  /**
   * The width of the media in pixels.
   * @type {number}
   * @memberof Media
   */
  width?: number;
}
// ...
export interface Place {
  /**
   *
   * @type {Array<string>}
   * @memberof Place
   */
  contained_within?: Array<string>;
  /**
   * The full name of the county in which this place exists.
   * @type {string}
   * @memberof Place
   */
  country?: string;
  /**
   * A two-letter ISO 3166-1 alpha-2 country code.
   * @type {string}
   * @memberof Place
   */
  country_code?: string;
  /**
   * The full name of this place.
   * @type {string}
   * @memberof Place
   */
  full_name: string;
  /**
   *
   * @type {Geo}
   * @memberof Place
   */
  geo?: Geo;
  /**
   * The identifier for this place.
   * @type {string}
   * @memberof Place
   */
  id: string;
  /**
   * The human readable name of this place.
   * @type {string}
   * @memberof Place
   */
  name?: string;
  /**
   *
   * @type {PlaceType}
   * @memberof Place
   */
  place_type?: PlaceType;
}
// ...
export interface Poll {
  /**
   *
   * @type {number}
   * @memberof Poll
   */
  duration_minutes?: number;
  /**
   *
   * @type {string}
   * @memberof Poll
   */
  end_datetime?: string;
  /**
   * Unique identifier of this poll.
   * @type {string}
   * @memberof Poll
   */
  id: string;
  /**
   *
   * @type {Array<PollOption>}
   * @memberof Poll
   */
  options: Array<PollOption>;
  /**
   *
   * @type {string}
   * @memberof Poll
   */
  voting_status?: PollVotingStatusEnum;
}
// ...
export interface Topic {
  /**
   * The description of the given topic.
   * @type {string}
   * @memberof Topic
   */
  description?: string;
  /**
   * Unique identifier of this Topic.
   * @type {string}
   * @memberof Topic
   */
  id: string;
  /**
   * The name of the given topic.
   * @type {string}
   * @memberof Topic
   */
  name: string;
}
/**
 *
 * @export
 * @interface Tweet
 */
export interface Tweet {
  /**
   *
   * @type {TweetAttachments}
   * @memberof Tweet
   */
  attachments?: TweetAttachments;
  /**
   * Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof Tweet
   */
  author_id?: string;
  /**
   *
   * @type {Array<ContextAnnotation>}
   * @memberof Tweet
   */
  context_annotations?: Array<ContextAnnotation>;
  /**
   * Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof Tweet
   */
  conversation_id?: string;
  /**
   * Creation time of the Tweet.
   * @type {string}
   * @memberof Tweet
   */
  created_at?: string;
  /**
   *
   * @type {FullTextEntities}
   * @memberof Tweet
   */
  entities?: FullTextEntities;
  /**
   *
   * @type {TweetGeo}
   * @memberof Tweet
   */
  geo?: TweetGeo;
  /**
   * Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof Tweet
   */
  id: string;
  /**
   * Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof Tweet
   */
  in_reply_to_user_id?: string;
  /**
   * Language of the Tweet, if detected by Twitter. Returned as a BCP47 language tag.
   * @type {string}
   * @memberof Tweet
   */
  lang?: string;
  /**
   *
   * @type {TweetNonPublicMetrics}
   * @memberof Tweet
   */
  non_public_metrics?: TweetNonPublicMetrics;
  /**
   *
   * @type {TweetOrganicMetrics}
   * @memberof Tweet
   */
  organic_metrics?: TweetOrganicMetrics;
  /**
   * Indicates if this Tweet contains URLs marked as sensitive, for example content suitable for mature audiences.
   * @type {boolean}
   * @memberof Tweet
   */
  possibly_sensitive?: boolean;
  /**
   *
   * @type {TweetPromotedMetrics}
   * @memberof Tweet
   */
  promoted_metrics?: TweetPromotedMetrics;
  /**
   *
   * @type {TweetPublicMetrics}
   * @memberof Tweet
   */
  public_metrics?: TweetPublicMetrics;
  /**
   * A list of Tweets this Tweet refers to. For example, if the parent Tweet is a Retweet, a Quoted Tweet or a Reply, it will include the related Tweet referenced to by its parent.
   * @type {Array<TweetReferencedTweetsInner>}
   * @memberof Tweet
   */
  referenced_tweets?: Array<TweetReferencedTweetsInner>;
  /**
   *
   * @type {ReplySettings}
   * @memberof Tweet
   */
  reply_settings?: ReplySettings;
  /**
   * The name of the app the user Tweeted from.
   * @type {string}
   * @memberof Tweet
   */
  source?: string;
  /**
   * The content of the Tweet.
   * @type {string}
   * @memberof Tweet
   */
  text: string;
  /**
   *
   * @type {TweetWithheld}
   * @memberof Tweet
   */
  withheld?: TweetWithheld;
}
/**
 * Specifies the type of attachments (if any) present in this Tweet.
 * @export
 * @interface TweetAttachments
 */
export interface TweetAttachments {
  /**
   * A list of Media Keys for each one of the media attachments (if media are attached).
   * @type {Array<string>}
   * @memberof TweetAttachments
   */
  media_keys?: Array<string>;
  /**
   * A list of poll IDs (if polls are attached).
   * @type {Array<string>}
   * @memberof TweetAttachments
   */
  poll_ids?: Array<string>;
}
// ...
export interface TweetGeo {
  /**
   *
   * @type {Point}
   * @memberof TweetGeo
   */
  coordinates?: Point;
  /**
   * The identifier for this place.
   * @type {string}
   * @memberof TweetGeo
   */
  place_id?: string;
}
// ...
export interface TweetNonPublicMetrics {
  /**
   * Number of times this Tweet has been viewed.
   * @type {number}
   * @memberof TweetNonPublicMetrics
   */
  impression_count?: number;
}
/**
 * Organic nonpublic engagement metrics for the Tweet at the time of the request.
 * @export
 * @interface TweetOrganicMetrics
 */
export interface TweetOrganicMetrics {
  /**
   * Number of times this Tweet has been viewed.
   * @type {number}
   * @memberof TweetOrganicMetrics
   */
  impression_count: number;
  /**
   * Number of times this Tweet has been liked.
   * @type {number}
   * @memberof TweetOrganicMetrics
   */
  like_count: number;
  /**
   * Number of times this Tweet has been replied to.
   * @type {number}
   * @memberof TweetOrganicMetrics
   */
  reply_count: number;
  /**
   * Number of times this Tweet has been Retweeted.
   * @type {number}
   * @memberof TweetOrganicMetrics
   */
  retweet_count: number;
}
/**
 * Promoted nonpublic engagement metrics for the Tweet at the time of the request.
 * @export
 * @interface TweetPromotedMetrics
 */
export interface TweetPromotedMetrics {
  /**
   * Number of times this Tweet has been viewed.
   * @type {number}
   * @memberof TweetPromotedMetrics
   */
  impression_count?: number;
  /**
   * Number of times this Tweet has been liked.
   * @type {number}
   * @memberof TweetPromotedMetrics
   */
  like_count?: number;
  /**
   * Number of times this Tweet has been replied to.
   * @type {number}
   * @memberof TweetPromotedMetrics
   */
  reply_count?: number;
  /**
   * Number of times this Tweet has been Retweeted.
   * @type {number}
   * @memberof TweetPromotedMetrics
   */
  retweet_count?: number;
}
/**
 * Engagement metrics for the Tweet at the time of the request.
 * @export
 * @interface TweetPublicMetrics
 */
export interface TweetPublicMetrics {
  /**
   * Number of times this Tweet has been liked.
   * @type {number}
   * @memberof TweetPublicMetrics
   */
  like_count: number;
  /**
   * Number of times this Tweet has been quoted.
   * @type {number}
   * @memberof TweetPublicMetrics
   */
  quote_count?: number;
  /**
   * Number of times this Tweet has been replied to.
   * @type {number}
   * @memberof TweetPublicMetrics
   */
  reply_count: number;
  /**
   * Number of times this Tweet has been Retweeted.
   * @type {number}
   * @memberof TweetPublicMetrics
   */
  retweet_count: number;
}
/**
 *
 * @export
 * @interface TweetReferencedTweetsInner
 */
export interface TweetReferencedTweetsInner {
  /**
   * Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof TweetReferencedTweetsInner
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof TweetReferencedTweetsInner
   */
  type: TweetReferencedTweetsInnerTypeEnum;
}
// ...
export interface User {
  /**
   * Creation time of this User.
   * @type {string}
   * @memberof User
   */
  created_at?: string;
  /**
   * The text of this User\'s profile description (also known as bio), if the User provided one.
   * @type {string}
   * @memberof User
   */
  description?: string;
  /**
   *
   * @type {UserEntities}
   * @memberof User
   */
  entities?: UserEntities;
  /**
   * Unique identifier of this User. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof User
   */
  id: string;
  /**
   * The location specified in the User\'s profile, if the User provided one. As this is a freeform value, it may not indicate a valid location, but it may be fuzzily evaluated when performing searches with location queries.
   * @type {string}
   * @memberof User
   */
  location?: string;
  /**
   * The friendly name of this User, as shown on their profile.
   * @type {string}
   * @memberof User
   */
  name: string;
  /**
   * Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
   * @type {string}
   * @memberof User
   */
  pinned_tweet_id?: string;
  /**
   * The URL to the profile image for this User.
   * @type {string}
   * @memberof User
   */
  profile_image_url?: string;
  /**
   * Indicates if this User has chosen to protect their Tweets (in other words, if this User\'s Tweets are private).
   * @type {boolean}
   * @memberof User
   */
  protected?: boolean;
  /**
   *
   * @type {UserPublicMetrics}
   * @memberof User
   */
  public_metrics?: UserPublicMetrics;
  /**
   * The URL specified in the User\'s profile.
   * @type {string}
   * @memberof User
   */
  url?: string;
  /**
   * The Twitter handle (screen name) of this user.
   * @type {string}
   * @memberof User
   */
  username: string;
  /**
   * Indicate if this User is a verified Twitter User.
   * @type {boolean}
   * @memberof User
   */
  verified?: boolean;
  /**
   *
   * @type {UserWithheld}
   * @memberof User
   */
  withheld?: UserWithheld;
}